+++ resources/js/components/modals/attribute/Edit.vue
@@ -72,6 +72,24 @@
                             >
                         </div>
                     </div>
+                    <div class="row">
+                        <label
+                            class="col-form-label text-end col-md-2"
+                            for="attribute-check-required"
+                        >
+                            {{ t('global.required') }}:
+                        </label>
+                        <div class="col-md-10 d-flex align-items-center">
+                            <div class="form-check form-switch">
+                                <input
+                                    id="attribute-check-required"
+                                    class="form-check-input"
+                                    type="checkbox"
+                                    v-model="state.required"
+                                >
+                            </div>
+                        </div>
+                    </div>
                 </template>
                 <hr>
                 <h5 class="text-center">
@@ -194,6 +212,13 @@
                         width: state.width,
                     };
                 }
+                if(state.attribute.pivot &&
+                    (!state.attribute.pivot.metadata || state.required != state.attribute.pivot.metadata.required)
+                ) {
+                    data.metadata = {
+                        required: state.required,
+                    };
+                }
                 if(state.attribute.is_system && state.separatorTitle && (!state.attribute.pivot.metadata?.title || state.separatorTitle != state.attribute.pivot.metadata.title)
                 ) {
                     data.metadata = {
@@ -231,6 +256,7 @@
                     groups: [getEmptyGroup(false)],
                 },
                 width: 100,
+                required: false,
                 isValid: computed(_ => {
                     return state.dependency.groups.every(group => {
                         return group.rules.length == 0 || group.rules.every(rule => validateDependencyRule(rule));
@@ -256,6 +282,7 @@
                 }
                 if(state.attribute?.pivot?.metadata) {
                     state.width = state.attribute.pivot.metadata.width || 100;
+                    state.required = state.attribute.pivot.metadata?.required === true;
                     if(state.attribute.is_system) {
                         state.separatorTitle = state.attribute.pivot.metadata.title;
                     }